Forms working, results.py page being called, result sets being filtered - just need to verify filter results are accurate, and then call table-generate
